Gulp tasks main refactoring
